Add AuthContext
Use ContextAPI to pass down authenticated state to multiple components; in a
more complicated codebase / project, might be worth using Redux, however in this
case, setting up ContextAPI is much simpler. The purpose of the AuthContext
would be to conditionally render pages or routes based on whether a user has
authenticated.